+++ Template/config/sidebar.php
@@ -1,3 +1,3 @@
-<li <?= $this->app->checkMenuSelection('ConfigController', 'show', ['plugin' => 'KanAI']) ?>>
+<li <?= $this->app->checkMenuSelection('ConfigController', 'show', 'KanAI') ?>>
     <?= $this->url->link(t('KanAI'), 'ConfigController', 'show', ['plugin' => 'KanAI']) ?>
 </li>